+++ frontend/ui/src/components/Stepper/GenericStepper.tsx
@@ -4,41 +4,15 @@ import { Box, Step } from "@mui/material";
 import Button from "@mui/material/Button";
 import Paper from "@mui/material/Paper";
 import Typography from "@mui/material/Typography";
-import { StepChildren } from "./StepChildren";
+import { StepChild } from "./StepChild";
 
-const steps = [
-  {
-    label: "Select campaign settings",
-    description: `For each ad campaign that you create, you can control how much
-              you're willing to spend on clicks and conversions, which networks
-              and geographical locations you want your ads to show on, and more.`,
-    content: <Typography variant="h4">This is FOURTH test</Typography>,
-  },
-  {
-    label: "Create an ad group",
-    description:
-      "An ad group contains one or more ads which target a shared set of keywords.",
-    content: <Typography variant="h4">This is a test</Typography>,
-  },
-  {
-    label: "Create an ad",
-    description: `Try out different ad text to see what brings in the most customers,
-              and learn how to enhance your ads using features like ad extensions.
-              If you run into any problems with your ads, find out how to tell if
-              they're running and how to resolve approval issues.`,
-    content: <Typography variant="h4">This is a test</Typography>,
-  },
-  {
-    label: "Create something",
-    description: `Try out different ad text to see what brings in the most customers,
-              and learn how to enhance your ads using features like ad extensions.
-              If you run into any problems with your ads, find out how to tell if
-              they're running and how to resolve approval issues.`,
-    content: <Typography variant="h4">This another test</Typography>,
-  },
-];
-
-export const GenericStepper: FC<{}> = ({}) => {
+export const GenericStepper: FC<{
+  steps: {
+    label: string;
+    description: string;
+    content: JSX.Element;
+  }[];
+}> = ({ steps }) => {
   const [activeStep, setActiveStep] = React.useState(0);
 
   const handleNext = () => {
@@ -58,7 +32,7 @@ export const GenericStepper: FC<{}> = ({}) => {
       <Stepper activeStep={activeStep} orientation="vertical">
         {steps.map((step, index) => (
           <Step key={step.label}>
-            <StepChildren
+            <StepChild
               step={step}
               handleNext={handleNext}
               handleBack={handleBack}
